+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "linkerd/http-metrics",
     "linkerd/http-retry",
     "linkerd/identity",
+    "linkerd/identity/default",
     "linkerd/io",
     "linkerd/metrics",
     "linkerd/opencensus",
@@ -37,8 +38,7 @@ members = [
     "linkerd/proxy/core",
     "linkerd/proxy/discover",
     "linkerd/proxy/http",
-    "linkerd/proxy/identity",
-    "linkerd/proxy/identity/default",
+    "linkerd/proxy/identity-client",
     "linkerd/proxy/resolve",
     "linkerd/proxy/tap",
     "linkerd/proxy/tcp",
