+++ linkerd/identity/src/lib.rs
@@ -1,81 +1,13 @@
 #![deny(warnings, rust_2018_idioms)]
 #![forbid(unsafe_code)]
 
+mod credentials;
+mod local;
+mod name;
+
+pub use self::{
+    credentials::{Credentials, DerX509},
+    local::LocalId,
+    name::Name,
+};
 pub use linkerd_dns_name::InvalidName;
-use std::{fmt, ops::Deref, str::FromStr, sync::Arc};
-
-/// An endpoint's identity.
-#[derive(Clone, Eq, PartialEq, Hash)]
-pub struct Name(Arc<linkerd_dns_name::Name>);
-
-/// A newtype for local server identities.
-#[derive(Clone, Debug, Eq, PartialEq, Hash)]
-pub struct LocalId(pub Name);
-
-// === impl Name ===
-
-impl From<linkerd_dns_name::Name> for Name {
-    fn from(n: linkerd_dns_name::Name) -> Self {
-        Name(Arc::new(n))
-    }
-}
-
-impl FromStr for Name {
-    type Err = InvalidName;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s.ends_with('.') {
-            return Err(InvalidName); // SNI hostnames are implicitly absolute.
-        }
-
-        linkerd_dns_name::Name::from_str(s).map(|n| Name(Arc::new(n)))
-    }
-}
-
-impl Deref for Name {
-    type Target = linkerd_dns_name::Name;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl fmt::Debug for Name {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        fmt::Debug::fmt(&self.0, f)
-    }
-}
-
-impl fmt::Display for Name {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        fmt::Display::fmt(&self.0, f)
-    }
-}
-
-impl From<LocalId> for Name {
-    fn from(LocalId(name): LocalId) -> Name {
-        name
-    }
-}
-
-// === impl LocalId ===
-
-impl From<Name> for LocalId {
-    fn from(n: Name) -> Self {
-        Self(n)
-    }
-}
-
-impl Deref for LocalId {
-    type Target = Name;
-
-    fn deref(&self) -> &Name {
-        &self.0
-    }
-}
-
-impl fmt::Display for LocalId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
