+++ linkerd/identity/src/name.rs
@@ -1,0 +1,46 @@
+use linkerd_dns_name::InvalidName;
+use std::{fmt, str::FromStr, ops::Deref, sync::Arc};
+
+/// An endpoint's identity.
+#[derive(Clone, Eq, PartialEq, Hash)]
+pub struct Name(Arc<linkerd_dns_name::Name>);
+
+// === impl Name ===
+
+impl From<linkerd_dns_name::Name> for Name {
+    fn from(n: linkerd_dns_name::Name) -> Self {
+        Name(Arc::new(n))
+    }
+}
+
+impl FromStr for Name {
+    type Err = InvalidName;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if s.ends_with('.') {
+            return Err(InvalidName); // SNI hostnames are implicitly absolute.
+        }
+
+        linkerd_dns_name::Name::from_str(s).map(|n| Name(Arc::new(n)))
+    }
+}
+
+impl Deref for Name {
+    type Target = linkerd_dns_name::Name;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl fmt::Debug for Name {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
+        fmt::Debug::fmt(&self.0, f)
+    }
+}
+
+impl fmt::Display for Name {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
+        fmt::Display::fmt(&self.0, f)
+    }
+}
