+++ linkerd/proxy/identity-client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "linkerd-proxy-identity"
+name = "linkerd-proxy-identity-client"
 version = "0.1.0"
 authors = ["Linkerd Developers <cncf-linkerd-dev@lists.cncf.io>"]
 license = "Apache-2.0"
