+++ linkerd/proxy/identity-client/src/certify.rs
@@ -1,7 +1,8 @@
-use crate::{Credentials, DerX509, Metrics, TokenSource};
+use crate::{Metrics, TokenSource};
 use http_body::Body;
 use linkerd2_proxy_api::identity::{self as api, identity_client::IdentityClient};
 use linkerd_error::{Error, Result};
+use linkerd_identity::{Credentials, DerX509};
 use linkerd_stack::NewService;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 use thiserror::Error;
