+++ linkerd/proxy/identity-client/src/lib.rs
@@ -1,0 +1,8 @@
+#![deny(warnings, rust_2018_idioms)]
+#![forbid(unsafe_code)]
+
+pub mod certify;
+pub mod metrics;
+mod token;
+
+pub use self::{certify::Certify, metrics::Metrics, token::TokenSource};
